Create Subplan During Template Creation (PR) (#99)
* Add a 'Create New Subplan' button with functionality in the program rules.

This allows users to create a subplan while they are in the middle of creating a program. Users will then be  able to save their subplan and return to their original form without losing data. If users accidentally click off the subplan form, or otherwise manage to close it, they are still able to navigate to the original form and have their changes persist. This PR closes issue #8